remove dependency on metrics package from engine/pkg/utils/kube/cache
