+++ controller/cache/mocks/LiveStateCache.go
@@ -5,7 +5,8 @@ package mocks
 import (
 	context "context"
 
-	metrics "github.com/argoproj/argo-cd/controller/metrics"
+	cache "github.com/argoproj/argo-cd/engine/pkg/utils/kube/cache"
+
 	kube "github.com/argoproj/argo-cd/engine/pkg/utils/kube"
 
 	mock "github.com/stretchr/testify/mock"
@@ -25,15 +26,15 @@ type LiveStateCache struct {
 }
 
 // GetClustersInfo provides a mock function with given fields:
-func (_m *LiveStateCache) GetClustersInfo() []metrics.ClusterInfo {
+func (_m *LiveStateCache) GetClustersInfo() []cache.ClusterInfo {
 	ret := _m.Called()
 
-	var r0 []metrics.ClusterInfo
-	if rf, ok := ret.Get(0).(func() []metrics.ClusterInfo); ok {
+	var r0 []cache.ClusterInfo
+	if rf, ok := ret.Get(0).(func() []cache.ClusterInfo); ok {
 		r0 = rf()
 	} else {
 		if ret.Get(0) != nil {
-			r0 = ret.Get(0).([]metrics.ClusterInfo)
+			r0 = ret.Get(0).([]cache.ClusterInfo)
 		}
 	}
 
@@ -114,11 +115,6 @@ func (_m *LiveStateCache) GetVersionsInfo(serverURL string) (string, []v1.APIGro
 	}
 
 	return r0, r1, r2
-}
-
-// Invalidate provides a mock function with given fields:
-func (_m *LiveStateCache) Invalidate() {
-	_m.Called()
 }
 
 // IsNamespaced provides a mock function with given fields: server, gk
