+++ engine/pkg/utils/kube/cache/cluster.go
@@ -21,7 +21,6 @@ import (
 	"k8s.io/client-go/rest"
 	"k8s.io/client-go/tools/pager"
 
-	"github.com/argoproj/argo-cd/controller/metrics"
 	"github.com/argoproj/argo-cd/engine/pkg/utils/health"
 	"github.com/argoproj/argo-cd/engine/pkg/utils/kube"
 )
@@ -36,6 +35,14 @@ type apiMeta struct {
 	namespaced      bool
 	resourceVersion string
 	watchCancel     context.CancelFunc
+}
+
+type ClusterInfo struct {
+	Server            string
+	K8SVersion        string
+	ResourcesCount    int
+	APIsCount         int
+	LastCacheSyncTime *time.Time
 }
 
 type Settings struct {
@@ -58,7 +65,7 @@ type ClusterCache interface {
 	IterateHierarchy(key kube.ResourceKey, action func(resource *Resource, namespaceResources map[kube.ResourceKey]*Resource))
 	IsNamespaced(gk schema.GroupKind) bool
 	GetManagedLiveObjs(targetObjs []*unstructured.Unstructured, isManaged func(r *Resource) bool) (map[kube.ResourceKey]*unstructured.Unstructured, error)
-	GetClusterInfo() metrics.ClusterInfo
+	GetClusterInfo() ClusterInfo
 }
 
 func NewClusterCache(settings Settings, config *rest.Config, namespaces []string, kubectl kube.Kubectl, handlers EventHandlers) *clusterCache {
@@ -684,10 +691,10 @@ var (
 	}
 )
 
-func (c *clusterCache) GetClusterInfo() metrics.ClusterInfo {
+func (c *clusterCache) GetClusterInfo() ClusterInfo {
 	c.lock.RLock()
 	defer c.lock.RUnlock()
-	return metrics.ClusterInfo{
+	return ClusterInfo{
 		APIsCount:         len(c.apisMeta),
 		K8SVersion:        c.serverVersion,
 		ResourcesCount:    len(c.resources),
